requirements.txt
# Web
fastapi>=0.115.0
hypercorn>=0.16.0

# Config & validation
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# DB
sqlalchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0

# Templates
jinja2>=3.1.4

# HTTP client
httpx>=0.27.0

# YAML
PyYAML>=6.0.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
asgi-lifespan>=2.1.0

aiosqlite>=0.20.0
pywebpush